feat(storage): добавлены репозитории для файлов и папок, обновлены миграции и модели для разделения имени и расширения файла; реализована сервисная логика загрузки с анонимным физическим хранением

+++ src/types/user.go
@@ -20,7 +20,7 @@ type Profile struct {
 type Account struct {
 	UserID       int       `json:"user_id" gorm:"primaryKey;column:user_id"`
 	Role         string    `json:"role" gorm:"column:role"`
-	StorageLimit int       `json:"storage_limit" gorm:"column:storage_limit"`
+	StorageLimit int64     `json:"storage_limit" gorm:"column:storage_limit"`
 	UpdatedAt    time.Time `json:"updated_at" gorm:"column:updated_at;autoUpdateTime"`
 }
